refactor: split provider config card into three blocks
- Break ProviderAccessSettingsBlock into ConfigureIdentitySettingsBlock,
  ConfigureConnectionSettingsBlock, ConfigureProtocolSettingsBlock
- Rename file to ConfigureSettingsBlocks.kt
- Reorder Block 2: Model ID -> Endpoint -> Toggle -> API Key
- Remove redundant onClearActiveField param (use fieldController directly)
- Restore deleted inline comments on protocol/thinking/deepseek compat
- Default vision model (supportsImages) to true for new configs
